+++ test/presentation/filter/pages/filter_create_edit_page_test.dart
@@ -60,6 +60,44 @@ void main() {
   group('FilterCreateEditPage', () {
     group('narrow view', () {
       group('create mode', () {
+        testWidgets('found no SaveFilterIconButton if name is empty',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(400, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(const BlocRepositoryWrapper());
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsNothing,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
+        testWidgets('found SaveFilterIconButton if name is not empty',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(400, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(const BlocRepositoryWrapper());
+          await tester.pumpAndSettle();
+          await tester.enterText(find.byType(TextFormField), 'Filter name');
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsOneWidget,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
         testWidgets('found no DeleteFilterIconButton', (tester) async {
           // Increase size to ensure all elements in list are visible.
           tester.view.physicalSize = const Size(400, 800);
@@ -79,6 +117,78 @@ void main() {
       });
 
       group('edit mode', () {
+        testWidgets(
+            'found no SaveFilterIconButton if filter has not be changed',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(400, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(
+            BlocRepositoryWrapper(
+              initFilter: const Filter().copyWith(name: 'filter'),
+            ),
+          );
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsNothing,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
+        testWidgets('found SaveFilterIconButton if filter has be changed',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(400, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(
+            BlocRepositoryWrapper(
+              initFilter: const Filter().copyWith(name: 'filter'),
+            ),
+          );
+          await tester.pumpAndSettle();
+          await tester.dragUntilVisible(
+            find.byType(FilterOrderItem),
+            find.byType(ListView),
+            const Offset(0, -100),
+          );
+          await tester.tap(find.byType(FilterOrderItem));
+          await tester.pumpAndSettle();
+          await tester.ensureVisible(find.byType(FilterStateOrderDialog));
+          await tester.pumpAndSettle();
+          await tester.tap(
+            find.descendant(
+              of: find.byType(FilterStateOrderDialog),
+              matching: find.text('Descending'),
+            ),
+          );
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(FilterOrderItem),
+              matching: find.byWidgetPredicate(
+                (Widget widget) =>
+                    widget is BasicChip &&
+                    widget.label == ListOrder.descending.name,
+              ),
+            ),
+            findsOneWidget,
+          );
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsOneWidget,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
         testWidgets('found DeleteFilterIconButton', (tester) async {
           // Increase size to ensure all elements in list are visible.
           tester.view.physicalSize = const Size(400, 800);
@@ -104,6 +214,44 @@ void main() {
 
     group('wide view', () {
       group('create mode', () {
+        testWidgets('found no SaveFilterIconButton if name is empty',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(800, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(const BlocRepositoryWrapper());
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsNothing,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
+        testWidgets('found SaveFilterIconButton if name is not empty',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(800, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(const BlocRepositoryWrapper());
+          await tester.pumpAndSettle();
+          await tester.enterText(find.byType(TextFormField), 'Filter name');
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsOneWidget,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
         testWidgets('found no DeleteFilterIconButton', (tester) async {
           // Increase size to ensure all elements in list are visible.
           tester.view.physicalSize = const Size(800, 800);
@@ -122,6 +270,78 @@ void main() {
         });
       });
       group('edit mode', () {
+        testWidgets(
+            'found no SaveFilterIconButton if filter has not be changed',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(800, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(
+            BlocRepositoryWrapper(
+              initFilter: const Filter().copyWith(name: 'filter'),
+            ),
+          );
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsNothing,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
+        testWidgets('found SaveFilterIconButton if filter has be changed',
+            (tester) async {
+          // Increase size to ensure all elements in list are visible.
+          tester.view.physicalSize = const Size(800, 800);
+          tester.view.devicePixelRatio = 1.0;
+          await tester.pumpWidget(
+            BlocRepositoryWrapper(
+              initFilter: const Filter().copyWith(name: 'filter'),
+            ),
+          );
+          await tester.pumpAndSettle();
+          await tester.dragUntilVisible(
+            find.byType(FilterOrderItem),
+            find.byType(ListView),
+            const Offset(0, -100),
+          );
+          await tester.tap(find.byType(FilterOrderItem));
+          await tester.pumpAndSettle();
+          await tester.ensureVisible(find.byType(FilterStateOrderDialog));
+          await tester.pumpAndSettle();
+          await tester.tap(
+            find.descendant(
+              of: find.byType(FilterStateOrderDialog),
+              matching: find.text('Descending'),
+            ),
+          );
+          await tester.pumpAndSettle();
+          expect(
+            find.descendant(
+              of: find.byType(FilterOrderItem),
+              matching: find.byWidgetPredicate(
+                (Widget widget) =>
+                    widget is BasicChip &&
+                    widget.label == ListOrder.descending.name,
+              ),
+            ),
+            findsOneWidget,
+          );
+          expect(
+            find.descendant(
+              of: find.byType(SaveFilterIconButton),
+              matching: find.byType(IconButton),
+            ),
+            findsOneWidget,
+          );
+          // resets the screen to its original size after the test end
+          addTearDown(tester.view.resetPhysicalSize);
+          addTearDown(tester.view.resetDevicePixelRatio);
+        });
         testWidgets('found DeleteFilterIconButton', (tester) async {
           // Increase size to ensure all elements in list are visible.
           tester.view.physicalSize = const Size(800, 800);